requirements.txt
# Data Engineering Dependencies
duckdb==0.9.2
pandas==2.1.4
python-dotenv==1.0.0
supabase==2.3.0
schedule==1.2.0
pyarrow==14.0.2
plotly==5.17.0
streamlit==1.29.0

